Refactor test fakes
